% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{make_fitness}
\alias{make_fitness}
\title{Making a fitness function for use in algorithms}
\usage{
make_fitness(function_, greater_is_better)
}
\arguments{
\item{function_}{A function to operate the fitness computation.}

\item{greater_is_better}{The direction of the fitness function.}
}
\value{
The list return includes the function itself and the sign of objective
}
\description{
make_fitness is used to create fitness function for user to apply in the algorithm.
}
\examples{
my_fitness <- function(x, y) {
  return(sum((x - y)^2))
}
my_customized_fitness <- make_fitness(my_fitness, FALSE)
my_customized_fitness
}
