% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function}
\alias{make_function}
\title{Making a function for use in programs}
\usage{
make_function(fun, arity, name, param = 0, param_pool = list())
}
\arguments{
\item{fun}{A function to operate the computation in execute algorithm.}

\item{arity}{The number of arguments that the function takes.}

\item{name}{The name of the function.}

\item{param}{The number of arguments that the function takes and varied during the training.}

\item{param_pool}{The possible values the param can take.}
}
\value{
The list return includes the function itself and the attributes of it.
}
\description{
make_function is used to create customized functions for user to apply in the algorithm.
}
\examples{
# define a function
my_function <- function(x, a, b) a * x^b
my_customized_function <- make_function(
  fun = my_function,
  name = "My first function",
  arity = 3,
  param = 2,
  param_pool = list(c(1, 2, 3), c(4, 5, 6))
)
# check the function return the correct number
my_customized_function$fun(5, 2, 3)
}
