% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic.R
\name{SymbolicRegressor}
\alias{SymbolicRegressor}
\title{A Genetic Programming symbolic regressor using R6 class.}
\description{
A symbolic regressor is an estimator that begins by building a population
of naive random formulas to represent a relationship.
The formulas are represented by tree-like structures with formulas as the nodes and variables as the leaves.
}
\details{
A regressor can fit the target values.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pop_size}}{The number of programs in each generation.}

\item{\code{generations}}{The number of generations to evolve.}

\item{\code{tournament_size}}{The number of programs that will compete to become part of the next generation.}

\item{\code{const_range}}{The range of constant can take for the program.}

\item{\code{init_depth}}{The range of tree depths for the initial population of naive formulas.}

\item{\code{init_method}}{The method define how the tree grows.}

\item{\code{function_set}}{A list of valid functions in list form to use in the program.}

\item{\code{metric}}{The raw fitness metric.}

\item{\code{parsimony_coefficient}}{The constant penalizes large programs by adjusting their fitness to be less favorable for selection.}

\item{\code{p_crossover}}{The probability of performing crossover on a tournament winner.}

\item{\code{p_subtree_mutation}}{The probability of performing subtree mutation on a tournament winner.}

\item{\code{p_hoist_mutation}}{The probability of performing hoist mutation on a tournament winner.}

\item{\code{p_point_mutation}}{The probability of performing point mutation on a tournament winner.}

\item{\code{p_point_replace}}{The probability that any given node will be mutated during point mutation.}

\item{\code{best_program}}{The flattened tree representation of the best program.}

\item{\code{feature_names}}{The vector of names of features of input data.}

\item{\code{method_probs}}{The probabilities to perform the crossover or mutations.}

\item{\code{programs}}{All the programs.}

\item{\code{fitness}}{Fitness of all the programs.
Initialize method}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SymbolicRegressor-new}{\code{SymbolicRegressor$new()}}
\item \href{#method-SymbolicRegressor-tournament}{\code{SymbolicRegressor$tournament()}}
\item \href{#method-SymbolicRegressor-evolve}{\code{SymbolicRegressor$evolve()}}
\item \href{#method-SymbolicRegressor-report_result}{\code{SymbolicRegressor$report_result()}}
\item \href{#method-SymbolicRegressor-fit}{\code{SymbolicRegressor$fit()}}
\item \href{#method-SymbolicRegressor-clone}{\code{SymbolicRegressor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolicRegressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolicRegressor-new}{}}}
\subsection{Method \code{new()}}{
Start a new regressor object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolicRegressor$new(
  pop_size = 2000,
  generations = 20,
  tournament_size = 20,
  const_range = c(-1, 1),
  init_depth = c(2:6),
  init_method = "half and half",
  function_set = list("add", "sub", "mul", "div"),
  metric = "mse",
  parsimony_coefficient = 0.001,
  p_crossover = 0.7,
  p_subtree_mutation = 0.1,
  p_hoist_mutation = 0.05,
  p_point_mutation = 0.1,
  p_point_replace = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop_size}}{The field pop_size.}

\item{\code{generations}}{The field generations.}

\item{\code{tournament_size}}{The field tournament_size.}

\item{\code{const_range}}{The field const_range.}

\item{\code{init_depth}}{The field init_depth.}

\item{\code{init_method}}{The field init_method.}

\item{\code{function_set}}{The field function_set.}

\item{\code{metric}}{The field metric.}

\item{\code{parsimony_coefficient}}{The field parsimony_coefficient.}

\item{\code{p_crossover}}{The field p_crossover.}

\item{\code{p_subtree_mutation}}{The field p_subtree_mutation.}

\item{\code{p_hoist_mutation}}{The field p_hoist_mutation.}

\item{\code{p_point_mutation}}{The field p_point_mutation.}

\item{\code{p_point_replace}}{The field p_point_replace.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new regressor object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolicRegressor-tournament"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolicRegressor-tournament}{}}}
\subsection{Method \code{tournament()}}{
Use tournament selection method to choose one program from the programs as the parent to evolve.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolicRegressor$tournament(programs, tournament_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{programs}}{The field programs.}

\item{\code{tournament_size}}{The field tournament_size.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A program.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolicRegressor-evolve"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolicRegressor-evolve}{}}}
\subsection{Method \code{evolve()}}{
Function used to build a batch of programs within a job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolicRegressor$evolve(parents, X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parents}}{List of programs}

\item{\code{X}}{The given feature data.}

\item{\code{y}}{The target values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The programs of current generation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolicRegressor-report_result"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolicRegressor-report_result}{}}}
\subsection{Method \code{report_result()}}{
A report of the progress of the evolution process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolicRegressor$report_result(gen, programs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gen}}{Current generation.}

\item{\code{programs}}{The programs of current generation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolicRegressor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolicRegressor-fit}{}}}
\subsection{Method \code{fit()}}{
The fit process of the estimator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolicRegressor$fit(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The feature data.}

\item{\code{y}}{The target values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SymbolicRegressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SymbolicRegressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SymbolicRegressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
